dafny: Implement a client for the all-new Dafny server
The client connects to the newly developed Dafny server, and persists
across an editing session; this lets Emacs take advantage of Dafny's
caching, instead of having to reverify the entire buffer after each
edit. This also means that Dafny's support in Emacs is (almost) on par
with Visual Studio.

The current default verification back-end is still the CLI, since the
server is experimental and has not yet been included in a release. There
is one pending issue that , which pertains to the tokens that Boogie
returns when an error is recycled (Boogie currently returns the original
tokens instead of the new ones. The VS extension compensates for this by
tracing snapshots and translating past buffer positions into new
ones. I'm not aware of a similar facility for emacs, unfortunately.

The protocol is documented in the header of inferior-dafny.el. Note that
this file, unlike the rest of this repo, is licensed under the MS-PL.