Update reserved keywords
Note, `least` and `greatest` are not reserved keywords, but `least predicate`, `greatest predicate`, `least lemma`, and `greatest lemma` are reserved keyword phrases. I listed these in the same way that `function method` and `predicate method` are listed in the file, even though I don’t understand the exact role of having a space in these strings in the definition of `dafny-defuns`.

There’s an infinite number of types `arrayN`, `arrayN?`, and `bvN`, where `N` is a number. (More precisely, `N` has no leading `0`’s unless `N` is `0`, and for arrays, `N` is not allowed to be `0` or `1`.) However, only the most common of these are represented as keywords here.