dafny/lsp: Add more known versions (up to 4.0.0)
This commit adds some more known versions to the automatic LSP setup.  There was
a packaging change as of version 3.13.0, so there is some conditional logic to
account for the difference.

All of the new versions were tested on Linux; the MacOS and Windows entries were
eyeballed based on the Github releases page.